logo,login and delete

+++ wealthwise-backend/src/main/java/com/wealthwise/wealthwise_backend/goal/repository/GoalRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 @Repository
 public interface GoalRepository extends JpaRepository<Goal, Integer> {
-    List<Goal> findByUserId(Integer userId);
+    List<Goal> findByUserId(Long userId);
 }
